Submit before revert

+++ PhoenixBot/UserAccountModel.cs
@@ -8,9 +8,9 @@ namespace PhoenixBot
     {
         public ulong Id { get; set; }
 
-        public int NumberOfWarnings { get; set; } = 0;
+        public byte NumberOfWarnings { get; set; } = 0;
 
-        public int IsMuted { get; set; } = 0;
+        public byte IsMuted { get; set; } = 0;
